requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.35
asyncpg==0.29.0
psycopg[binary]==3.2.1
apscheduler==3.10.4
python-dotenv==1.0.1
